requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
aiosqlite
python-jose[cryptography]
passlib
python-multipart
pytest
httpx